requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pdfplumber==0.11.4
faiss-cpu
sentence-transformers==2.7.0
InstructorEmbedding==1.0.1
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118 
transformers
accelerate

